buddy allocator r/w

+++ kernel/src/memory/physical/allocator.rs
@@ -1,8 +1,8 @@
 pub trait MemoryAllocator : Send + Sync {
     fn new() -> Self;
 
-    fn init(&mut self, start: u64, end: u64);
-    fn data_size(&self) -> u64;
+    fn init(&mut self, start: usize, end: usize);
+    fn data_size(&self) -> usize;
     fn set_data(&mut self, data: *mut u8);
-    fn include(&mut self, start: u64, end: u64);
+    fn include(&mut self, start: usize, end: usize);
 }
